[#167][delivery] refactor: 패키지 이동
