+++ delivery/src/main/java/com/faster/delivery/app/delivery/application/MessageClient.java
@@ -1,6 +1,6 @@
 package com.faster.delivery.app.delivery.application;
 
-import com.faster.delivery.app.deliverymanager.application.dto.SendMessageApplicationRequestDto;
+import com.faster.delivery.app.delivery.application.dto.SendMessageApplicationRequestDto;
 import org.springframework.web.bind.annotation.RequestBody;
 
 public interface MessageClient {
