+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/client/MessageClientImpl.java
@@ -1,8 +1,8 @@
 package com.faster.delivery.app.delivery.infrastructure.client;
 
 import com.faster.delivery.app.delivery.application.MessageClient;
-import com.faster.delivery.app.deliverymanager.application.dto.SendMessageApplicationRequestDto;
-import com.faster.delivery.app.deliverymanager.infrastructure.client.dto.SendMessageFeignRequestDto;
+import com.faster.delivery.app.delivery.application.dto.SendMessageApplicationRequestDto;
+import com.faster.delivery.app.delivery.infrastructure.client.dto.SendMessageFeignRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -18,3 +18,7 @@ public class MessageClientImpl implements MessageClient {
         SendMessageFeignRequestDto.from(dto));
   }
 }
+
+/**
+ * 아이디어 경로에 배송담당자 배송 가능시간 ~때부터는 배송이 가능하다
+  */
