+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/client/MessageFeignClient.java
@@ -1,7 +1,7 @@
 package com.faster.delivery.app.delivery.infrastructure.client;
 
 import com.common.response.ApiResponse;
-import com.faster.delivery.app.deliverymanager.infrastructure.client.dto.SendMessageFeignRequestDto;
+import com.faster.delivery.app.delivery.infrastructure.client.dto.SendMessageFeignRequestDto;
 import com.faster.delivery.app.global.config.FeignClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
